Extract a few test helpers
